Rename instance -> executorrepository and fix index.js exports
