+++ cloud/wfm/src/executor/ExecutorRepository.ts
@@ -1,0 +1,6 @@
+import ProcessInstance from '../process-instance/ProcessInstance';
+interface ExecutorRepository {
+  saveInstance(instance: ProcessInstance): Promise<ProcessInstance>;
+}
+
+export default ExecutorRepository;
