--- cloud/wfm/src/executor/InstanceRepository.ts
@@ -1,6 +1,0 @@
-import ProcessInstance from '../process-instance/ProcessInstance';
-interface InstanceRepository {
-  save(instance: ProcessInstance): Promise<ProcessInstance>;
-}
-
-export default InstanceRepository;
