+++ cloud/wfm/test/Executor/ExecutorImpl.ts
@@ -1,13 +1,13 @@
 import * as assert from 'assert';
 import * as Promise from 'bluebird';
 import ExecutorImpl from '../../src/executor/ExecutorImpl';
-import InstanceRepository from '../../src/executor/InstanceRepository';
+import ExecutorRepository from '../../src/executor/InstanceRepository';
 import ProcessInstance from '../../src/process-instance/ProcessInstance';
 import suite from './index';
 
-class SingleRepository implements InstanceRepository {
+class SingleRepository implements ExecutorRepository {
   public savedInstance: ProcessInstance;
-  public save(instance: ProcessInstance) {
+  public saveInstance(instance: ProcessInstance) {
     this.savedInstance = instance;
     return Promise.resolve(instance);
   }
